:lipstick: 스타일 분리 및 keypress 속성 추가
